ZavaManufacturing-ISA95/mqtt-simulator/requirements.txt
paho-mqtt>=2.0.0,<3.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
PyYAML>=6.0,<7.0
